Added identifier and execution status getters for micro-ops
